move Gpt2Generator
