+++ autobestof_cleanup.py
@@ -1,4 +1,4 @@
-import gpt2_config
+from gpt2_config import Gpt2Generator
 from reddit_auth import reddit
 import wordpress
 
@@ -30,39 +30,13 @@ def write_wordpress_post(comment):
     resp = wordpress.create_post(wp_post)
     return resp.status_code
 
-class Gpt2Generator(object):
-    def __init__(self, max_tokens=915):
-        self.max_tokens = max_tokens
-        self.encoder = gpt2_config.get_encoder()
-        self.sess, self.gpt2 = gpt2_config.load_gpt2()
-
-    def generate_titles(self, comment):
-        try:
-            author = comment.author.name
-        except:
-            author = 'u/Redditor'
-        body_ = '\n'.join([author, comment.body])
-        max_body_tokens = self.max_tokens - 15  # we will append more tokens
-        body_trunc = self.encoder.decode(self.encoder.encode(body_)[:max_body_tokens])
-        body = '\n'.join([body_trunc, comment.id, '<|rbestof|>'])
-        titles = self.gpt2.generate(self.sess,
-            length=100,
-            temperature=0.7,
-            prefix=body,
-            nsamples=5,
-            batch_size=5,
-            include_prefix=False,
-            truncate='<|endoftext|>',
-            return_as_list=True
-        )
-        titles = [title.split(maxsplit=1)[1] for title in titles]
-        return titles
-
 
 if __name__== '__main__':
     r_autobestof = reddit.subreddit('autobestof')
     wp_posts = [post['title']['rendered'] for post in wordpress.get_posts()]
     gpt2 = Gpt2Generator()
+    gpt2.get_encoder()
+    gpt2.load_gpt2()
     for submission in r_autobestof.stream.submissions():
         sub_url = submission.url.split('/')
         link_id = sub_url[6]
